split decoding functionality
